+++ source/CMakeROOT_6_14.cmake
@@ -120,6 +120,7 @@ add_library(ostap SHARED src/format.cpp
                          src/NStatEntity.cpp
                          src/Notifier.cpp
                          src/OstapDataFrame.cpp
+                         src/P2Quantile.cpp
                          src/Parameterization.cpp
                          src/Params.cpp
                          src/Peaks.cpp