+++ source/src/StatVar.cpp
@@ -24,6 +24,7 @@
 #include "Ostap/MatrixUtils.h"
 #include "Ostap/StatVar.h"
 #include "Ostap/FormulaVar.h"
+#include "Ostap/P2Quantile.h"
 // ============================================================================
 // Local
 // ============================================================================
@@ -561,7 +562,7 @@ namespace
    *   @param  last  (INPUT) the last event to  process
    *   @return the quantile value 
    */
-  std::vector<double> 
+  Ostap::StatVar::Quantiles
   _quantiles_
   ( TTree&                  tree      ,
     const std::set<double>& quantiles , //  0<q<1 
@@ -592,7 +593,7 @@ namespace
       ++num ;
     }
     //
-    if ( 0 == num ) { return std::vector<double>() ; }
+    if ( 0 == num ) { return Ostap::StatVar::Quantiles ( std::vector<double>() , num ) ; }
     //
     typedef std::vector<double> VALUES ;
     VALUES values{} ; values.reserve ( num ) ;
@@ -612,6 +613,7 @@ namespace
       //
       var.evaluate  ( results ) ;
       values.insert ( values.end() , results.begin() , results.end() ) ;
+      //
     }
     //
     std::vector<double> result ; result.reserve ( quantiles.size() ) ;
@@ -625,7 +627,7 @@ namespace
       result.push_back ( *start ) ;
     }
     //
-    return result ;
+    return Ostap::StatVar::Quantiles ( result , values.size () ) ; 
   }
   // ==========================================================================
   /*   get quantile of the distribution  
@@ -637,7 +639,55 @@ namespace
    *   @param  last  (INPUT) the last event to  process
    *   @return the quantile value 
    */
-  std::vector<double> 
+  Ostap::StatVar::Quantiles
+  _p2quantiles_
+  ( TTree&                  tree      ,
+    const std::set<double>& quantiles , //  0<q<1 
+    Ostap::Formula&         var       ,
+    Ostap::Formula*         cuts      , 
+    const unsigned long     first     ,
+    const unsigned long     last      ) 
+  {
+    // the loop 
+    const unsigned long the_last = std::min ( last , (unsigned long) tree.GetEntries() ) ;
+    //
+    Ostap::Utils::Notifier notify ( &tree , &var , cuts ) ;
+    const bool with_cuts = nullptr != cuts ? true : false ;
+    //
+    std::vector<Ostap::Math::GSL::P2Quantile> qs ( quantiles.begin() , quantiles.end() ) ;
+    //
+    unsigned long num = 0  ;
+    std::vector<double> results {} ;
+    for ( unsigned long entry = first ; entry < the_last ; ++entry ) 
+    {      
+      long ievent = tree.GetEntryNumber ( entry ) ;
+      if ( 0 > ievent ) { break ; }                        // BREAK
+      //
+      ievent      = tree.LoadTree ( ievent ) ;
+      if ( 0 > ievent ) { break ; }                        // BREAK
+      //
+      const long double w = with_cuts ? cuts->evaluate() : 1.0L ;
+      //
+      if ( !w  ) { continue ; }                           // CONTINUE       
+      //
+      var.evaluate  ( results ) ;
+      for ( auto& q : qs ) { q.add  ( results.begin() , results.end () ) ; }
+      num += results.size();
+    }
+    //
+    return Ostap::StatVar::Quantiles ( std::vector<double>( qs.begin (), qs.end () )  , num ) ;
+  }
+  // ==========================================================================
+  /*   get quantile of the distribution  
+   *   @param tree  (INPUT) the input tree 
+   *   @param q     (INPUT) quantile value   0 < q < 1  
+   *   @param expr  (INPUT) the expression 
+   *   @param cuts  (INPUT) selection cuts 
+   *   @param  first (INPUT) the first  event to process 
+   *   @param  last  (INPUT) the last event to  process
+   *   @return the quantile value 
+   */
+  Ostap::StatVar::Quantiles 
   _quantiles_
   ( const RooAbsData&       data      ,
     const std::set<double>& quantiles , //  0<q<1 
@@ -705,7 +755,57 @@ namespace
       result.push_back ( *start ) ;
     }
     //
-    return result ;
+    return Ostap::StatVar::Quantiles ( result , values.size () ) ; 
+  }
+  // ==========================================================================
+  /*   get (Appeoximate) quantile of the distribution using  P^2 algorithm  
+   *   @param tree  (INPUT) the input tree 
+   *   @param q     (INPUT) quantile value   0 < q < 1  
+   *   @param expr  (INPUT) the expression 
+   *   @param cuts  (INPUT) selection cuts 
+   *   @param  first (INPUT) the first  event to process 
+   *   @param  last  (INPUT) the last event to  process
+   *   @return the quantile value 
+   */
+  Ostap::StatVar::Quantiles 
+  _p2quantiles_
+  ( const RooAbsData&       data      ,
+    const std::set<double>& quantiles , //  0<q<1 
+    const RooAbsReal&       var       ,
+    const RooAbsReal*       cuts      , 
+    const unsigned long     first     ,
+    const unsigned long     last      , 
+    const char*             cut_range ) 
+  {
+    // the loop 
+    const unsigned long the_last = std::min ( last , (unsigned long) data.numEntries() ) ;
+    //
+    const bool  weighted = data.isWeighted () ;
+    //
+    std::vector<Ostap::Math::GSL::P2Quantile> qs ( quantiles.begin() , quantiles.end() ) ;
+    //
+    unsigned long num = 0 ;
+    for ( unsigned long entry = first ; entry < the_last ; ++entry )
+    {
+      const RooArgSet* vars = data.get( entry ) ;
+      if ( nullptr == vars )                              { break    ; } // BREAK 
+      //
+      if ( cut_range && !vars->allInRange ( cut_range ) ) { continue ; } // CONTINUE    
+      // apply cuts:
+      const long double wc = nullptr != cuts ? cuts -> getVal() : 1.0L ;
+      if ( !wc ) { continue ; }                                          // CONTINUE  
+      // apply weight:
+      const long double wd = weighted  ? data.weight()   : 1.0L ;
+      if ( !wd ) { continue ; }                                          // CONTINUE    
+      // cuts & weight:
+      const long double w  = wd *  wc ; 
+      if ( !w  ) { continue ; }                                          // CONTINUE        
+      //
+      for ( auto& q : qs ) { q.add ( var.getVal() ) ; }
+      ++num ;
+    }
+    //
+    return Ostap::StatVar::Quantiles ( std::vector<double>( qs.begin (), qs.end () )  , num ) ;
   }
   // ==========================================================================
 } //                                                 end of anonymous namespace
@@ -1683,7 +1783,8 @@ Ostap::StatVar::kurtosis
  *   @return the quantile value 
  */
 // ============================================================================
-double Ostap::StatVar::quantile
+Ostap::StatVar::Quantile
+Ostap::StatVar::quantile
 ( TTree&              tree  ,
   const double        q     , //  0<q<1 
   const std::string&  expr  , 
@@ -1709,15 +1810,65 @@ double Ostap::StatVar::quantile
                     "Ostap::StatVar::quantile"     ) ;
   }
   //
-  const std::vector<double> result = _quantiles_ 
+  auto result = _quantiles_ 
     ( tree , std::set<double> {{ q }} , var , cut.get() , first , last ) ; 
   //
-  Ostap::Assert ( 1 == result.size()         , 
+  Ostap::Assert ( 1 == result.quantiles.size()         , 
                   "Invalid quantiles size"   ,
                   "Ostap::StatVar::interval" ) ;
   //
-  return result[0] ;
+  return Quantile ( result.quantiles[0] , result.nevents ) ;
 }
+// ============================================================================
+/*   get approximate  quantile of the distribution  using p^2 algortihm
+ *   @param tree  (INPUT) the input tree 
+ *   @param q     (INPUT) quantile value   0 < q < 1  
+ *   @param expr  (INPUT) the expression 
+ *   @param cuts  (INPUT) selection cuts 
+ *   @param  first (INPUT) the first  event to process 
+ *   @param  last  (INPUT) the last event to  process
+ *   @return the quantile value 
+ */
+// ============================================================================
+Ostap::StatVar::Quantile
+Ostap::StatVar::p2quantile
+( TTree&              tree  ,
+  const double        q     , //  0<q<1 
+  const std::string&  expr  , 
+  const std::string&  cuts  , 
+  const unsigned long first ,
+  const unsigned long last  ) 
+{
+  Ostap::Assert ( 0 < q && q < 1             , 
+                  "Invalid quantile"         ,
+                  "Ostap::StatVar::quantile" ) ;
+  //
+  Ostap::Formula var ( expr , &tree ) ;
+  Ostap::Assert ( var.ok()                              , 
+                  "Invalid expression:\"" + expr + "\"" ,
+                  "Ostap::StatVar::quantile"            ) ;
+  //
+  std::unique_ptr<Ostap::Formula> cut { nullptr } ;
+  if  ( !cuts.empty() ) 
+  { 
+    cut = std::make_unique<Ostap::Formula>( "", cuts , &tree ) ; 
+    Ostap::Assert ( cut && cut->ok()               ,
+                    "Invalid cut:\"" + cuts + "\"" ,
+                    "Ostap::StatVar::quantile"     ) ;
+  }
+  //
+  auto result = _p2quantiles_ 
+    ( tree , std::set<double> {{ q }} , var , cut.get() , first , last ) ; 
+  //
+  Ostap::Assert ( 1 == result.quantiles.size()         , 
+                  "Invalid quantiles size"   ,
+                  "Ostap::StatVar::interval" ) ;
+  //
+  return Quantile ( result.quantiles[0] , result.nevents ) ;
+}
+// ============================================================================
+
+
 // ============================================================================
 /*   get quantiles of the distribution  
  *   @param tree  (INPUT) the input tree 
@@ -1729,7 +1880,7 @@ double Ostap::StatVar::quantile
  *   @return the quantile value 
  */
 // ============================================================================
-std::vector<double> 
+Ostap::StatVar::Quantiles
 Ostap::StatVar::quantiles
 ( TTree&                     tree      ,
   const std::vector<double>& quantiles , 
@@ -1768,6 +1919,55 @@ Ostap::StatVar::quantiles
   return _quantiles_ ( tree , qs , var  ,  cut.get() , first , last ) ; 
 }
 // ============================================================================
+/*   get (approximate) quantiles of the distribution using P^2 algorithm   
+ *   @param tree  (INPUT) the input tree 
+ *   @param q     (INPUT) quantile value   0 < q < 1  
+ *   @param expr  (INPUT) the expression 
+ *   @param cuts  (INPUT) selection cuts 
+ *   @param  first (INPUT) the first  event to process 
+ *   @param  last  (INPUT) the last event to  process
+ *   @return the quantile value 
+ */
+// ============================================================================
+Ostap::StatVar::Quantiles
+Ostap::StatVar::p2quantiles
+( TTree&                     tree      ,
+  const std::vector<double>& quantiles , 
+  const std::string&         expr      , 
+  const std::string&         cuts      , 
+  const unsigned long        first     ,
+  const unsigned long        last      ) 
+{
+  //
+  std::set<double> qs ;
+  for ( double v : quantiles ) { qs.insert ( v ) ; }
+  Ostap::Assert ( !qs.empty ()                ,
+                  "Invalid quantiles"         ,
+                  "Ostap::StatVar::quantiles" ) ;
+  Ostap::Assert ( 0 < *qs. begin ()           , 
+                  "Invalid quantile"          ,
+                  "Ostap::StatVar::quantiles" ) ;  
+  Ostap::Assert ( 1 > *qs.rbegin ()           , 
+                  "Invalid quantile"          ,
+                  "Ostap::StatVar::quantiles" ) ;
+  //
+  Ostap::Formula var ( expr , &tree ) ;
+  Ostap::Assert ( var.ok()                              ,
+                  "Invalid expression:\"" + expr + "\"" ,
+                  "Ostap::StatVar::quantile"            ) ;
+  //
+  std::unique_ptr<Ostap::Formula> cut { nullptr } ;
+  if  ( !cuts.empty() ) 
+  { 
+    cut = std::make_unique<Ostap::Formula>( cuts , &tree ) ; 
+    Ostap::Assert ( cut && cut->ok()               , 
+                    "Invalid cut:\"" + cuts + "\"" ,
+                    "Ostap::StatVar::quantile"     ) ;
+  }
+  //
+  return _p2quantiles_ ( tree , qs , var  ,  cut.get() , first , last ) ; 
+}
+// ============================================================================
 /*  get the interval of the distribution  
  *   @param tree  (INPUT) the input tree 
  *   @param q1    (INPUT) quantile value   0 < q1 < 1  
@@ -1784,7 +1984,7 @@ Ostap::StatVar::quantiles
  *   @code 
  */
 // ============================================================================
-Ostap::StatVar::Interval 
+Ostap::StatVar::QInterval 
 Ostap::StatVar::interval 
 ( TTree&              tree  ,
   const double        q1    , //  0<q1<1 
@@ -1815,14 +2015,71 @@ Ostap::StatVar::interval
                     "Ostap::StatVar::interval"     ) ;
   }
   //
-  const std::vector<double> result = 
+  auto result = 
     _quantiles_ ( tree , std::set<double>{{ q1 , q2 }} , 
                   var  ,  cut.get() , first , last ) ; 
-  Ostap::Assert ( 2 == result.size()         ,
+  Ostap::Assert ( 2 == result.quantiles.size()         ,
                   "Invalid interval"         ,
                   "Ostap::StatVar::interval" ) ;
   //
-  return std::make_pair( result[0] , result[1] ) ;
+  return QInterval ( Interval ( result.quantiles[0] , result.quantiles[1] ) , result.nevents ) ;
+}
+// ============================================================================
+/*  get the approximate  interval of the distribution  usnig P^2 algorithm
+ *   @param tree  (INPUT) the input tree 
+ *   @param q1    (INPUT) quantile value   0 < q1 < 1  
+ *   @param q2    (INPUT) quantile value   0 < q2 < 1  
+ *   @param expr  (INPUT) the expression 
+ *   @param cuts  (INPUT) selection cuts 
+ *   @param  first (INPUT) the first  event to process 
+ *   @param  last  (INPUT) the last event to  process
+ *   @return the quantile value 
+ *   @code
+ *   Tree& tree = ... ;
+ *   /// get 90% interval:
+ *   Interval ab = p2interval ( tree , 0.05 , 0.95 , 'mass' , 'pt>3' ) ;
+ *   @code 
+ */
+// ============================================================================
+Ostap::StatVar::QInterval 
+Ostap::StatVar::p2interval 
+( TTree&              tree  ,
+  const double        q1    , //  0<q1<1 
+  const double        q2    , //  0<q2<1 
+  const std::string&  expr  , 
+  const std::string&  cuts  , 
+  const unsigned long first ,
+  const unsigned long last  ) 
+{
+  Ostap::Assert ( 0 < q1 && q1 < 1 , 
+                  "Invalid quantile1"        ,
+                  "Ostap::StatVar::interval" ) ;
+  Ostap::Assert ( 0 < q2 && q2 < 1 , 
+                  "Invalid quantile2"        ,
+                  "Ostap::StatVar::interval" ) ;
+  //
+  Ostap::Formula var ( expr , &tree ) ;
+  Ostap::Assert ( var.ok()                              , 
+                  "Invalid expression:\"" + expr + "\"" ,
+                  "Ostap::StatVar::interval"            ) ;
+  //
+  std::unique_ptr<Ostap::Formula> cut { nullptr } ;
+  if  ( !cuts.empty() ) 
+  { 
+    cut = std::make_unique<Ostap::Formula>( "", cuts , &tree ) ; 
+    Ostap::Assert ( cut && cut->ok()               ,
+                    "Invalid cut:\"" + cuts + "\"" ,
+                    "Ostap::StatVar::interval"     ) ;
+  }
+  //
+  auto result = 
+    _p2quantiles_ ( tree , std::set<double>{{ q1 , q2 }} , 
+                    var  ,  cut.get() , first , last ) ; 
+  Ostap::Assert ( 2 == result.quantiles.size()         ,
+                  "Invalid interval"         ,
+                  "Ostap::StatVar::interval" ) ;
+  //
+  return QInterval ( Interval ( result.quantiles[0] , result.quantiles[1] ) , result.nevents ) ;
 }
 // ============================================================================
 /** get the number of equivalent entries 
@@ -2300,7 +2557,8 @@ Ostap::StatVar::kurtosis
  *   @return the quantile value 
  */
 // ============================================================================
-double Ostap::StatVar::quantile
+Ostap::StatVar::Quantile
+Ostap::StatVar::quantile
 ( const RooAbsData&   data      ,
   const double        q         , //  0<q<1 
   const std::string&  expr      , 
@@ -2322,15 +2580,60 @@ double Ostap::StatVar::quantile
   const std::unique_ptr<Ostap::FormulaVar> expression { make_formula ( expr , data        ) } ;
   const std::unique_ptr<Ostap::FormulaVar> cut        { make_formula ( cuts , data , true ) } ;
   //  
-  const std::vector<double> result = _quantiles_ 
+  auto result = _quantiles_ 
     (  data,  std::set<double>{{ q }} , 
        *expression ,  cut.get() , first , the_last , cutrange ) ;
   //
-  Ostap::Assert ( 1 == result.size()         ,
+  Ostap::Assert ( 1 == result.quantiles.size()         ,
                   "Invalid quantile size"    ,
                   "Ostap::StatVar::quantile" ) ;
   //
-  return result[0] ;
+  return Quantile ( result.quantiles[0] , result.nevents ) ;
+}
+// ============================================================================
+/*   get (approximate) quantile of the distribution using P^2 algorithm
+ *   @param data   (INPUT) the input data
+ *   @param q      (INPUT) quantile value   0 < q < 1  
+ *   @param expr   (INPUT) the expression 
+ *   @param cuts   (INPUT) selection cuts 
+ *   @param  first (INPUT) the first  event to process 
+ *   @param  last  (INPUT) the last event to  process
+ *   @param  cut_range (INPUT) cut range 
+ *   @return the quantile value 
+ */
+// ============================================================================
+Ostap::StatVar::Quantile
+Ostap::StatVar::p2quantile
+( const RooAbsData&   data      ,
+  const double        q         , //  0<q<1 
+  const std::string&  expr      , 
+  const std::string&  cuts      , 
+  const std::string&  cut_range , 
+  const unsigned long first     ,
+  const unsigned long last      )
+{
+  Ostap::Assert ( 0 < q && q < 1             , 
+                  "Invalid quantile"         ,
+                  "Ostap::StatVar::quantile" ) ;
+  //
+  const unsigned long num_entries = data.numEntries() ;
+  const unsigned long the_last    = std::min ( num_entries , last ) ;
+  if ( the_last <= first ) { return  0 ; }    // RETURN
+  //
+  const char* cutrange  = cut_range.empty() ?  nullptr : cut_range.c_str() ;
+  //
+  const std::unique_ptr<Ostap::FormulaVar> expression { make_formula ( expr , data        ) } ;
+  const std::unique_ptr<Ostap::FormulaVar> cut        { make_formula ( cuts , data , true ) } ;
+  //  
+  auto result = _p2quantiles_ 
+    (  data,  std::set<double>{{ q }} , 
+       *expression ,  cut.get() , first , the_last , cutrange ) ;
+  //
+  Ostap::Assert ( 1 == result.quantiles.size()         ,
+                  "Invalid quantile size"    ,
+                  "Ostap::StatVar::quantile" ) ;
+  //
+  return Quantile ( result.quantiles[0] , result.nevents ) ;
 }
 // ============================================================================
 /*   get the interval of the distribution  
@@ -2349,7 +2652,7 @@ double Ostap::StatVar::quantile
  *   @code 
  */
 // ============================================================================
-Ostap::StatVar::Interval 
+Ostap::StatVar::QInterval 
 Ostap::StatVar::interval 
 ( const RooAbsData&   data      ,
   const double        q1        , //  0<q1<1 
@@ -2369,21 +2672,74 @@ Ostap::StatVar::interval
   //
   const unsigned long num_entries = data.numEntries() ;
   const unsigned long the_last    = std::min ( num_entries , last ) ;
-  if ( the_last <= first ) { return  std::make_pair(0.0,0.0) ; }    // RETURN
+  if ( the_last <= first ) { return QInterval() ; }    // RETURN
   //
   const char* cutrange  = cut_range.empty() ?  nullptr : cut_range.c_str() ;
   //
   const std::unique_ptr<Ostap::FormulaVar> expression { make_formula ( expr , data        ) } ;
   const std::unique_ptr<Ostap::FormulaVar> cut        { make_formula ( cuts , data , true ) } ;
   //  
-  const std::vector<double> result = _quantiles_ 
+  auto result = _quantiles_ 
     (  data,  std::set<double>{{ q1 , q2 }} , 
        *expression ,  cut.get() , first , the_last , cutrange ) ;
-  Ostap::Assert ( 2 ==  result.size()        ,
+  Ostap::Assert ( 2 ==  result.quantiles.size()        ,
                   "Invalid quantile size"    ,
                   "Ostap::StatVar::quantile" ) ;
   //
-  return std::make_pair( result[0] , result[1] ) ;
+  return QInterval ( Interval ( result.quantiles[0] , result.quantiles [1] ) , result.nevents ) ;
+}
+// ============================================================================
+/*   get the approximate  interval of the distribution  usnig  P^2 algorithm
+ *   @param data  (INPUT) the input data
+ *   @param q1    (INPUT) quantile value   0 < q1 < 1  
+ *   @param q2    (INPUT) quantile value   0 < q2 < 1  
+ *   @param expr  (INPUT) the expression 
+ *   @param cuts  (INPUT) selection cuts 
+ *   @param  first (INPUT) the first  event to process 
+ *   @param  last  (INPUT) the last event to  process
+ *   @return the quantile value 
+ *   @code
+ *   const RooAbsData& data = ... ;
+ *   /// get 90% interval:
+ *   Interval ab = p2interval ( data , 0.05 , 0.95 , 'mass' , 'pt>3' ) ;
+ *   @code 
+ */
+// ============================================================================
+Ostap::StatVar::QInterval 
+Ostap::StatVar::p2interval 
+( const RooAbsData&   data      ,
+  const double        q1        , //  0<q1<1 
+  const double        q2        , //  0<q2<1 
+  const std::string&  expr      , 
+  const std::string&  cuts      , 
+  const std::string&  cut_range , 
+  const unsigned long first     ,
+  const unsigned long last      )
+{
+  Ostap::Assert ( 0 < q1 && q1 < 1           , 
+                  "Invalid quantile1"        ,
+                  "Ostap::StatVar::quantile" ) ;
+  Ostap::Assert ( 0 < q2 && q2 < 1           , 
+                  "Invalid quantile2"        ,
+                  "Ostap::StatVar::quantile" ) ;
+  //
+  const unsigned long num_entries = data.numEntries() ;
+  const unsigned long the_last    = std::min ( num_entries , last ) ;
+  if ( the_last <= first ) { return  QInterval()  ; }    // RETURN
+  //
+  const char* cutrange  = cut_range.empty() ?  nullptr : cut_range.c_str() ;
+  //
+  const std::unique_ptr<Ostap::FormulaVar> expression { make_formula ( expr , data        ) } ;
+  const std::unique_ptr<Ostap::FormulaVar> cut        { make_formula ( cuts , data , true ) } ;
+  //  
+  auto result = _p2quantiles_ 
+    (  data,  std::set<double>{{ q1 , q2 }} , 
+       *expression ,  cut.get() , first , the_last , cutrange ) ;
+  Ostap::Assert ( 2 ==  result.quantiles.size()        ,
+                  "Invalid quantile size"    ,
+                  "Ostap::StatVar::quantile" ) ;
+  //
+  return QInterval ( Interval ( result.quantiles[0] , result.quantiles [1] ) , result.nevents ) ;
 }
 // ============================================================================
 /*   get quantiles of the distribution  
@@ -2396,7 +2752,7 @@ Ostap::StatVar::interval
  *   @return the quantile value 
  */
 // ============================================================================
-std::vector<double> 
+Ostap::StatVar::Quantiles
 Ostap::StatVar::quantiles
 ( const RooAbsData&          data      ,
   const std::vector<double>& quantiles , 
@@ -2430,6 +2786,53 @@ Ostap::StatVar::quantiles
   return _quantiles_ ( data  , qs  , 
                        *expression , cut.get() , 
                        first , the_last , cutrange ) ;
+}
+// ============================================================================
+/*   get (approximate) quantiles of the distribution using P^2 algorithm  
+ *   @param data  (INPUT) the input data
+ *   @param q     (INPUT) quantile value   0 < q < 1  
+ *   @param expr  (INPUT) the expression 
+ *   @param cuts  (INPUT) selection cuts 
+ *   @param  first (INPUT) the first  event to process 
+ *   @param  last  (INPUT) the last event to  process
+ *   @return the quantile value 
+ */
+// ============================================================================
+Ostap::StatVar::Quantiles
+Ostap::StatVar::p2quantiles
+( const RooAbsData&          data      ,
+  const std::vector<double>& quantiles , 
+  const std::string&         expr      , 
+  const std::string&         cuts      , 
+  const std::string&         cut_range , 
+  const unsigned long        first     ,
+  const unsigned long        last      )
+{
+  std::set<double> qs ;
+  for ( double v : quantiles ) { qs.insert( v ) ; }
+  Ostap::Assert (  !qs.empty()                 , 
+                   "Invalid quantiles"         ,
+                   "Ostap::StatVar::quantiles" ) ;
+  Ostap::Assert ( 0 < *qs. begin ()            ,  
+                  "Invalid quantile"           ,
+                  "Ostap::StatVar::quantiles"  ) ;
+  Ostap::Assert ( 1 > *qs.rbegin ()            ,  
+                  "Invalid quantile"           ,
+                  "Ostap::StatVar::quantiles"  ) ;
+  //
+  const unsigned long num_entries = data.numEntries() ;
+  const unsigned long the_last    = std::min ( num_entries , last ) ;
+  if ( the_last <= first )
+  { return  Ostap::StatVar::Quantiles  ( std::vector<double>() , 0 ) ; }    // RETURN
+  //
+  const char* cutrange  = cut_range.empty() ?  nullptr : cut_range.c_str() ;
+  //
+  const std::unique_ptr<Ostap::FormulaVar> expression { make_formula ( expr , data        ) } ;
+  const std::unique_ptr<Ostap::FormulaVar> cut        { make_formula ( cuts , data , true ) } ;
+  //  
+  return _p2quantiles_ ( data  , qs  , 
+                         *expression , cut.get() , 
+                         first , the_last , cutrange ) ;
 }
 // ============================================================================
 // Actions with frames 
@@ -3015,8 +3418,8 @@ Ostap::Math::ValueWithError Ostap::StatVar::kurtosis
 namespace 
 {
   // ==========================================================================
-  /// get quantiles 
-  std::vector<double> _quantiles_ 
+  /// get quantiles
+  Ostap::StatVar::Quantiles _quantiles_ 
   ( Ostap::DataFrame        frame , 
     const std::set<double>& qs    , 
     const std::string&      expr  , 
@@ -3047,9 +3450,32 @@ namespace
       start = values.begin() + current ;
       result.push_back ( *start ) ;
     }
-    values.clear() ;
-    
-    return  result ;
+    return  Ostap::StatVar::Quantiles ( result , values.size() ) ;
+  }
+  // ==========================================================================
+  /// get approximate  quantiles using P^2 algorithm  
+  Ostap::StatVar::Quantiles
+  _p2quantiles_ 
+  ( Ostap::DataFrame        frame     , 
+    const std::set<double>& quantiles , 
+    const std::string&      expr      , 
+    const std::string&      cuts      ) 
+  {
+    const bool no_cuts = trivial (  cuts ) ; 
+    //
+    const std::string var   = Ostap::tmp_name ( "v_"   , expr ) ;
+    const std::string bcut  = Ostap::tmp_name ( "b_"   , cuts ) ;
+    //
+    std::vector<Ostap::Math::GSL::P2Quantile> qs ( quantiles.begin() , quantiles.end() ) ;
+    // decorate the frame 
+    auto t = frame
+      .Define   ( bcut    , no_cuts ? "true" : "(bool) ( " + cuts + " ) ;" ) 
+      .Filter   ( bcut    )    
+      .Define   ( var     , "1.0*(" + expr + ")" ) ;
+    auto l = t.Count() ;
+    t.Foreach  (  [&qs] ( double v ) { for  ( auto&q : qs ) { q.add ( v ) ; } } , { var } ) ;
+    // 
+    return  Ostap::StatVar::Quantiles ( std::vector<double>( qs.begin() , qs.end() ) , *l ) ;
   }
   // ==========================================================================
 }
@@ -3062,7 +3488,8 @@ namespace
  *   @return the quantile value 
  */
 // ============================================================================
-double Ostap::StatVar::quantile
+Ostap::StatVar::Quantile
+Ostap::StatVar::quantile
 ( Ostap::DataFrame    frame  ,
   const double        q      , //  0<q<1 
   const std::string&  expr   , 
@@ -3072,10 +3499,35 @@ double Ostap::StatVar::quantile
                   "Invalid quantile"         ,
                   "Ostap::StatVar::quantile" ) ;
   auto result = _quantiles_ ( frame , std::set<double>{{ q }} , expr , cuts ) ;
-  Ostap::Assert ( 1 == result.size()         , 
+  Ostap::Assert ( 1 == result.quantiles.size()         , 
                   "Invalid quantiles size"   ,
                   "Ostap::StatVar::interval" ) ;
-  return result [0] ;
+  return Ostap::StatVar::Quantile ( result.quantiles[0] , result.nevents ) ;
+}
+// ============================================================================
+/**  get approximate quantile of the distribution using  P^2 algorithm  
+ *   @param frame  (INPUT) the input data
+ *   @param q      (INPUT) quantile value   0 < q < 1  
+ *   @param expr   (INPUT) the expression 
+ *   @param cuts   (INPUT) selection cuts 
+ *   @return the quantile value 
+ */
+// ============================================================================
+Ostap::StatVar::Quantile
+Ostap::StatVar::p2quantile
+( Ostap::DataFrame    frame  ,
+  const double        q      , //  0<q<1 
+  const std::string&  expr   , 
+  const std::string&  cuts   ) 
+{
+  Ostap::Assert ( 0 < q && q < 1             , 
+                  "Invalid quantile"         ,
+                  "Ostap::StatVar::quantile" ) ;
+  auto result = _p2quantiles_ ( frame , std::set<double>{{ q }} , expr , cuts ) ;
+  Ostap::Assert ( 1 == result.quantiles.size()         , 
+                  "Invalid quantiles size"   ,
+                  "Ostap::StatVar::interval" ) ;
+  return Ostap::StatVar::Quantile ( result.quantiles[0] , result. nevents ) ;
 }
 // ========================================================================    
 /*  get quantiles of the distribution  
@@ -3086,7 +3538,8 @@ double Ostap::StatVar::quantile
  *   @return the quantile value 
  */
 // ========================================================================    
-std::vector<double> Ostap::StatVar::quantiles
+Ostap::StatVar::Quantiles
+Ostap::StatVar::quantiles
 ( Ostap::DataFrame           frame     ,
   const std::vector<double>& quantiles , 
   const std::string&         expr      , 
@@ -3109,6 +3562,39 @@ std::vector<double> Ostap::StatVar::quantiles
   //
   return _quantiles_ ( frame , qs , expr , cuts ) ;
 }
+// ========================================================================    
+/*   get approximate  quantiles of the distribution  using P^2 algorithm 
+ *   @param frame (INPUT) the input frame
+ *   @param q     (INPUT) quantile value   0 < q < 1  
+ *   @param expr  (INPUT) the expression 
+ *   @param cuts  (INPUT) selection cuts 
+ *   @return the quantile value 
+ */
+// ========================================================================    
+Ostap::StatVar::Quantiles
+Ostap::StatVar::p2quantiles
+( Ostap::DataFrame           frame     ,
+  const std::vector<double>& quantiles , 
+  const std::string&         expr      , 
+  const std::string&         cuts      ) 
+{
+  Ostap::Assert ( 1 <= quantiles.size()          , 
+                  "Invalid vector of quantiles"  ,
+                  "Ostap::StatVar::quantile"     ) ;
+  std::set<double> qs ;
+  for ( double v : quantiles ) { qs.insert ( v ) ; }
+  Ostap::Assert ( !qs.empty ()                ,
+                  "Invalid quantiles"         ,
+                  "Ostap::StatVar::quantiles" ) ;
+  Ostap::Assert ( 0 < *qs. begin ()           , 
+                  "Invalid quantile"          ,
+                  "Ostap::StatVar::quantiles" ) ;  
+  Ostap::Assert ( 1 > *qs.rbegin ()           , 
+                  "Invalid quantile"          ,
+                  "Ostap::StatVar::quantiles" ) ;
+  //
+  return _p2quantiles_ ( frame , qs , expr , cuts ) ;
+}
 // ============================================================================
 /* Get the interval of the distribution  
  * @param tree  (INPUT) the input tree 
@@ -3124,7 +3610,7 @@ std::vector<double> Ostap::StatVar::quantiles
  * @code 
  */
 // ============================================================================
-Ostap::StatVar::Interval 
+Ostap::StatVar::QInterval 
 Ostap::StatVar::interval 
 ( Ostap::DataFrame    frame ,
   const double        q1    , //  0<q1<1 
@@ -3139,13 +3625,49 @@ Ostap::StatVar::interval
                   "Invalid quantile2"        ,
                   "Ostap::StatVar::interval" ) ;
   //
-  const std::vector<double> result = 
-    _quantiles_ ( frame , std::set<double>{{ q1 , q2 }} ,  expr , cuts ) ; 
-  Ostap::Assert ( 2 == result.size()         ,
+  auto result = _quantiles_( frame , std::set<double>{{ q1 , q2 }} ,  expr , cuts ) ; 
+  Ostap::Assert ( 2 == result.quantiles.size()         ,
                   "Invalid interval"         ,
                   "Ostap::StatVar::interval" ) ;
   //
-  return std::make_pair( result[0] , result[1] ) ;
+  return QInterval ( Interval ( result.quantiles[0] , result.quantiles[1] ) , result.nevents ) ;
+}
+// ============================================================================
+/* Get the approximate  interval of the distribution  using P^2 algorithm
+ * @param tree  (INPUT) the input tree 
+ * @param q1    (INPUT) quantile value   0 < q1 < 1  
+ * @param q2    (INPUT) quantile value   0 < q2 < 1  
+ * @param expr  (INPUT) the expression 
+ * @param cuts  (INPUT) selection cuts 
+ * @return the quantile value 
+ * @code
+ * FRAME& frame = ... ;
+ * /// get 90% interval:
+ * Interval ab = interval ( frame , 0.05 , 0.95 , 'mass' , 'pt>3' ) ;
+ * @code 
+ */
+// =======================x=====================================================
+Ostap::StatVar::QInterval 
+Ostap::StatVar::p2interval 
+( Ostap::DataFrame    frame ,
+  const double        q1    , //  0<q1<1 
+  const double        q2    , //  0<q2<1 
+  const std::string&  expr  , 
+  const std::string&  cuts  )
+{
+  Ostap::Assert ( 0 < q1 && q1 < 1 , 
+                  "Invalid quantile1"        ,
+                  "Ostap::StatVar::interval" ) ;
+  Ostap::Assert ( 0 < q2 && q2 < 1 , 
+                  "Invalid quantile2"        ,
+                  "Ostap::StatVar::interval" ) ;
+  //
+  auto result = _p2quantiles_ ( frame , std::set<double>{{ q1 , q2 }} ,  expr , cuts ) ; 
+  Ostap::Assert ( 2 == result.quantiles.size()         ,
+                  "Invalid interval"         ,
+                  "Ostap::StatVar::interval" ) ;
+  //
+  return QInterval ( Interval ( result.quantiles[0] , result.quantiles[1] ) , result.nevents ) ;
 }
 // ============================================================================
 // The END